First cut of a hand-modified BUILD.gn file for GRPC.
In a previous CL we added BUILD.gn.template and used it to
generate a BUILD.gn file but we did not hand-edit the BUILD.gn file
and we did not try to make it work. In this CL we take a first
cut at hand-editing it so that GRPC compiles.

This allows GRPC to build and we are able to run a unit test that does
not actually use the network. We have not yet tried to actually use the
network. One open question is what to do with the dependency on c-ares.
Currently we have removed it from the build and that seems ok so far.

Change-Id: Ib1bf57e54491b85792740dcc1432eb88ed088990