Adds BUILD.gn.template and uses it to generate BUILD.gn.
We copied BUILD.gn.template from
https://chromium.googlesource.com/external/github.com/grpc/grpc/+log/chromium-deps/2016-10-25/templates/BUILD.gn.template

We then ran tools/buildgen/generate_projects.sh and discarded
all changes other than the newly generated BUILD.gn file.

This BUILD.gn file is not quite perfect for Fuchsia, but we are
committing it as a base against which we will later make some
manual changes.

Change-Id: Ibcea54bbaf133b8bd073c438eb5029ae3b362fe2