Modify template to sort list of source files alphabetically.
This is another step on the way to
https://fuchsia-review.googlesource.com/#/c/33945/.

I want to make it clear which source files have been
added or deleted to the automatically generated
BUILD.gn file in my manual edit. In this CL I modify the
template in order to make cosmetic-only changes to BUILD.gn.
That way in the following CL it will be easier to see the
substantive changes.
(1) Sort the list of source files in each source_set
(2) Delete some unneeded source_sets.
(3) Delete some Chromium-specific configs.

Change-Id: Id48ef59b327a1b777646f499d8030f7980c92147